chore: moved scroll percentage function to shared
