+++ packages/shared/src/utility.ts
@@ -10,8 +10,33 @@ function debounce(callback: TimerHandler, frequency = 500, timer = 0) {
 
 /* -----------------------------------
  *
+ * ScrollPercentage
+ *
+ * -------------------------------- */
+
+function getScrollPercentage() {
+  const body = document.body;
+  const scrollTop = window.pageYOffset || body.scrollTop;
+  const { scrollHeight, offsetHeight, clientHeight } = document.documentElement;
+
+  const documentHeight = Math.max(
+    body.scrollHeight,
+    scrollHeight,
+    body.offsetHeight,
+    offsetHeight,
+    body.clientHeight,
+    clientHeight
+  );
+
+  const trackLength = documentHeight - window.innerHeight;
+
+  return Math.floor(Math.abs(scrollTop / trackLength) * 100);
+}
+
+/* -----------------------------------
+ *
  * Export
  *
  * -------------------------------- */
 
-export { debounce };
+export { debounce, getScrollPercentage };
